[Misc] Consolidate Apple platform stubs
